% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_test_delta_sweep.R
\name{gmm_test_delta_sweep}
\alias{gmm_test_delta_sweep}
\title{Compute gamma}
\usage{
gmm_test_delta_sweep(X, R, delta_max, k, n_deltas = 100, compute_gamma = TRUE)
}
\arguments{
\item{X}{A numeric vector.}

\item{R}{A numeric vector.}

\item{delta_max}{Maximum for the range.}

\item{k}{A numeric parameter.}

\item{n_deltas}{A number of steps.}

\item{compute_gamma}{A logical value}
}
\value{
The index of the first sign change, or NA if no sign change occurs.
}
\description{
Compute gamma
}
