% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_test_homoskedasticity_robust.R
\name{gmm_test_homoskedasticity_robust}
\alias{gmm_test_homoskedasticity_robust}
\title{Robust version of GMM test for the delta sweep}
\usage{
gmm_test_homoskedasticity_robust(E, S, regularization = 1e-08)
}
\arguments{
\item{E}{A numeric vector.}

\item{S}{A numeric vector.}

\item{regularization}{a tolerance}
}
\value{
The index of the first sign change, or NA if no sign change occurs.
}
\description{
Robust version of GMM test for the delta sweep
}
