% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msiv_reg.R
\name{msiv_reg}
\alias{msiv_reg}
\title{This function performs Instrumental Variable Regression using multiple synthetic IVs}
\usage{
msiv_reg(data, Y, X, H, method = "simple", reps = 5)
}
\arguments{
\item{data}{A data frame.}

\item{Y}{Name of the dependent variable.}

\item{X}{Character vector of endogenous variable names.}

\item{H}{Character vector of exogenous variable names.}

\item{method}{Either "simple" or "GMM". Simple method is based on cov(e^2,siv)=0. GMM method is based on J=M'WM=0.}

\item{reps}{Number of bootstrap loops.}
}
\value{
A list containing IV regression results and generated SIVs.
}
\description{
The synthetic IVs are identified using the dual tendency DT condition for coplanar regression vectors.
The DT is formulated for both the homoscedastic and heteroscedastic cases.
}
\examples{
df <- wooldridge::mroz
data <- df[complete.cases(df), ]
result <- msiv_reg(data, "hours", c("lwage", "educ"),
                   c("age", "kidslt6", "kidsge6", "nwifeinc"), method="simple", reps=5)
result$IV1
result$IV2
result$IV3
result1 <- msiv_reg(data, "hours", c("lwage", "educ"),
                   c("age", "kidslt6", "kidsge6", "nwifeinc"), method="GMM", reps=5)
result1$IV1
}
